+++ mkdocs.yml
@@ -16,6 +16,7 @@ plugins:
                     show_object_full_path: True
                     show_source: False  # If true, it would be hard to use Ctrl+F to search in the doc
                                         # (it will search in the code!)
+                    docstring_section_style: list
         watch:
             - python/
             # We cannot watch build/. Reload mkdocs manually after rebuilding
